Agno(Phidata) /Day 02/requirements.txt
agno
ollama
groq
python-dotenv
duckduckgo-search
streamlit
yfinance
plotly